+++ lib/screens/main_screen.dart
@@ -37,9 +37,9 @@ import 'package:flutterlifecyclehooks/flutterlifecyclehooks.dart';
 import 'package:package_info/package_info.dart';
 import 'package:pi_authenticator_legacy/pi_authenticator_legacy.dart';
 import 'package:privacyidea_authenticator/model/tokens.dart';
-import 'package:privacyidea_authenticator/screens/add_manually_screen.dart';
+import 'package:privacyidea_authenticator/screens/add_token_screen.dart';
 import 'package:privacyidea_authenticator/screens/changelog_screen.dart';
-import 'package:privacyidea_authenticator/screens/custom_about_screen.dart';
+import 'package:privacyidea_authenticator/screens/about_screen.dart';
 import 'package:privacyidea_authenticator/screens/guide_screen.dart';
 import 'package:privacyidea_authenticator/screens/scanner_screen.dart';
 import 'package:privacyidea_authenticator/screens/settings_screen.dart';
